app/requirements.txt
asttokens==2.2.1
av==10.0.0
backcall==0.2.0
boto3==1.26.118
botocore==1.29.118
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
coloredlogs==15.0.1
ctranslate2==3.12.0
decorator==5.1.1
einops==0.6.1
encodec==0.1.1
executing==1.2.0
faster-whisper==0.4.1
filelock==3.12.0
Flask==2.2.3
Flask-Cors==3.0.10
flatbuffers==23.3.3
funcy==2.0
huggingface-hub==0.13.4
humanfriendly==10.0
idna==3.4
ipython==8.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
lit==16.0.1
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
onnxruntime==1.14.1
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.38
protobuf==4.22.3
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
python-dateutil==2.8.2
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
s3transfer==0.6.0
scipy==1.10.1
six==1.16.0
stack-data==0.6.2
suno-bark @ git+https://github.com/suno-ai/bark.git@97c6019ecdf1a70fdb313835c3bb3f770a0459e4
sympy==1.11.1
tokenizers==0.13.3
torch==2.0.0
torchaudio==2.0.1
tqdm==4.65.0
traitlets==5.9.0
transformers==4.28.1
triton==2.0.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.2.3
